added background to main
Added dice rolls for abilities and moved code around in better order.
